Created two exceptions, added exceptions to Animal method move(int speed), refactored code to java google style
