+++ src/ru/task3/zoo/actions/Huntable.java
@@ -1,4 +1,4 @@
-package ru.task3.actions;
+package ru.task3.zoo.actions;
 
 public interface Huntable {
     public void hunt();
