+++ src/ru/task3/zoo/exception/HugeSpeedException.java
@@ -1,0 +1,7 @@
+package ru.task3.zoo.exception;
+
+public class HugeSpeedException extends Exception{
+    public HugeSpeedException(String message) {
+        super(message);
+    }
+}
