+++ src/ru/task3/zoo/exception/NegativeSpeedException.java
@@ -1,0 +1,7 @@
+package ru.task3.zoo.exception;
+
+public class NegativeSpeedException extends Exception{
+    public NegativeSpeedException(String message) {
+        super(message);
+    }
+}
